Separate player endpoint
